Add densenet169, densenet210 models
